Loaded GDT, have IDT stub. Need to redirect PIC output most likely and properly configure TSS. Then can continue to paging/virtual memory

+++ lib/k_string.c
@@ -30,7 +30,7 @@ int k_strcpy(char* dst, const char* src)
   while(*src)
     *dst++ = *src++;
   *dst = 0; // NULL terminate
-  return start;
+  return (int)start;
 }
 
 void k_memset(char* dst, int val, int len)
